@@ -20,69 +20,67 @@
         "algolia/algoliasearch-client-php": "^3.3",
         "doctrine/dbal": "^3.5",
         "drewm/mailchimp-api": "^2.5",
-        "filament/filament": "^2.16",
-        "filament/spatie-laravel-tags-plugin": "^2.0",
+        "filament/filament": "^3.2",
+        "filament/spatie-laravel-tags-plugin": "^3.2",
         "guzzlehttp/guzzle": "^7.4.5",
         "knplabs/github-api": "dev-master#9b01208e1e5500201b45699c911ad53d08f59105",
-        "laravel/framework": "^10.48",
-        "laravel/horizon": "^5.12",
-        "laravel/sail": "^1.18",
-        "laravel/sanctum": "^3.3",
-        "laravel/slack-notification-channel": "^2.5",
-        "laravel/telescope": "^4.12",
-        "laravel/tinker": "^2.8",
-        "laravel/ui": "^4.2",
-        "livewire/livewire": "^2.11",
+        "laravel/framework": "^11.0",
+        "laravel/horizon": "^5.22",
+        "laravel/sail": "^1.26",
+        "laravel/sanctum": "^4.0",
+        "laravel/slack-notification-channel": "^3.2",
+        "laravel/telescope": "^5.0",
+        "laravel/tinker": "^2.9",
+        "laravel/ui": "^4.4",
+        "livewire/livewire": "^3.3",
         "meilisearch/meilisearch-php": "0.20",
         "ohdearapp/ohdear-php-sdk": "^3.4.4",
         "predis/predis": "^1.1.10",
-        "spatie/browsershot": "^3.57",
+        "spatie/browsershot": "^4.0",
         "spatie/commonmark-highlighter": "^3.0",
         "spatie/cpu-load-health-check": "^1.0.2",
-        "spatie/filament-markdown-editor": "^1.0",
+        "spatie/filament-markdown-editor": "dev-main",
         "spatie/fork": "^1.1",
-        "spatie/laravel-backup": "^8.1.2",
-        "spatie/laravel-comments-livewire": "^1.2.5",
-        "spatie/laravel-feed": "^4.2",
-        "spatie/laravel-flash": "^1.9",
-        "spatie/laravel-health": "^1.18",
-        "spatie/laravel-honeypot": "^4.3",
-        "spatie/laravel-horizon-watcher": "^1.0",
-        "spatie/laravel-ignition": "^2.0",
-        "spatie/laravel-log-dumper": "^1.4.1",
+        "spatie/laravel-backup": "^8.6",
+        "spatie/laravel-comments-livewire": "^2.0",
+        "spatie/laravel-feed": "^4.4",
+        "spatie/laravel-flash": "^1.10",
+        "spatie/laravel-health": "^1.27",
+        "spatie/laravel-honeypot": "^4.5",
+        "spatie/laravel-horizon-watcher": "^1.1",
+        "spatie/laravel-ignition": "^2.4",
+        "spatie/laravel-log-dumper": "dev-main",
         "spatie/laravel-mailcoach-mailer": "^1.0",
         "spatie/laravel-mailcoach-sdk": "^1.0",
-        "spatie/laravel-markdown": "^2.2.4",
-        "spatie/laravel-medialibrary": "^10.7",
-        "spatie/laravel-menu": "^4.1",
-        "spatie/laravel-missing-page-redirector": "^2.9.2",
-        "spatie/laravel-model-info": "^1.4",
-        "spatie/laravel-ray": "^1.32",
-        "spatie/laravel-remote": "^1.3",
-        "spatie/laravel-responsecache": "^7.4",
+        "spatie/laravel-markdown": "^2.4",
+        "spatie/laravel-medialibrary": "^11.3",
+        "spatie/laravel-menu": "^4.2",
+        "spatie/laravel-model-info": "^2.0",
+        "spatie/laravel-ray": "^1.35",
+        "spatie/laravel-remote": "^1.4",
+        "spatie/laravel-responsecache": "^7.5",
         "spatie/laravel-route-discovery": "^1.0.1",
-        "spatie/laravel-schedule-monitor": "^3.1",
-        "spatie/laravel-site-search": "^1.4",
-        "spatie/laravel-stubs": "^2.5",
-        "spatie/laravel-tags": "^4.3.2",
-        "spatie/laravel-tail": "^4.4",
-        "spatie/laravel-webhook-client": "^3.1.2",
+        "spatie/laravel-schedule-monitor": "^3.7",
+        "spatie/laravel-site-search": "^2.2",
+        "spatie/laravel-tags": "^4.6",
+        "spatie/laravel-tail": "^4.5",
+        "spatie/laravel-webhook-client": "^3.3",
         "spatie/security-advisories-health-check": "^0.0.2",
-        "spatie/sheets": "^1.10",
+        "spatie/sheets": "^1.12",
         "spatie/valuestore": "^1.3",
         "stripe/stripe-php": "^7.128",
-        "symfony/http-client": "^6.2",
-        "symfony/mailgun-mailer": "^6.2",
-        "symfony/postmark-mailer": "^6.2",
+        "symfony/http-client": "^7.0",
+        "symfony/mailgun-mailer": "^7.0",
+        "symfony/postmark-mailer": "^7.0",
         "wyrihaximus/react-child-process-promise": "^3.0"
     },
     "require-dev": {
-        "barryvdh/laravel-ide-helper": "^2.13",
-        "brianium/paratest": "^6.11",
-        "fakerphp/faker": "^1.20",
-        "mockery/mockery": "^1.5",
-        "nunomaduro/collision": "^6.2.1",
-        "phpunit/phpunit": "^9.5.21"
+        "barryvdh/laravel-ide-helper": "^3.0",
+        "brianium/paratest": "^7.0",
+        "fakerphp/faker": "^1.23",
+        "mockery/mockery": "^1.6",
+        "nunomaduro/collision": "^8.1",
+        "phpunit/phpunit": "^10.5"
     },
     "config": {
         "optimize-autoloader": true,
