@@ -11,7 +11,10 @@ final class Talk extends Model
     use HasFactory;
     use TalkPresenter;
 
-    protected $casts = [
-        'presented_at' => 'datetime',
-    ];
+    protected function casts(): array
+    {
+        return [
+            'presented_at' => 'datetime',
+        ];
+    }
 }
