@@ -8,9 +8,9 @@ use App\Filament\Resources\PostResource\Pages;
 use App\Modules\Blog\Models\Post;
 use Filament\Forms;
 use Filament\Forms\Components\SpatieTagsInput;
-use Filament\Resources\Form;
+use Filament\Forms\Form;
 use Filament\Resources\Resource;
-use Filament\Resources\Table;
+use Filament\Tables\Table;
 use Filament\Tables;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
