@@ -8,9 +8,9 @@ use App\Filament\Resources\SponsorsResource\CreateSponsor;
 use App\Filament\Resources\SponsorsResource\EditSponsor;
 use App\Filament\Resources\SponsorsResource\ListSponsors;
 use Filament\Forms;
-use Filament\Resources\Form;
+use Filament\Forms\Form;
 use Filament\Resources\Resource;
-use Filament\Resources\Table;
+use Filament\Tables\Table;
 use Filament\Tables;
 use Livewire\TemporaryUploadedFile;
 
